Improve student UI feedback and support decimal grades
Enhanced ConsoleUI to provide colored feedback for errors and successful actions, improved input validation for grades, and updated grade handling to support decimal values. Student.cs was updated to store grades as doubles instead of integers for more precise grading.

+++ EduTrack/Student.cs
@@ -7,17 +7,17 @@ namespace EduTrack
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public int Grade { get; set; }
-        public Dictionary<string, List<int>> Subjects { get; set; } = new();
+        public Dictionary<string, List<double>> Subjects { get; set; } = new();
 
         public override string ToString()
         {
             return $"{FirstName} {LastName} | Grade: {Grade}";
         }
 
-        public void AddGrade(string subject, int grade)
+        public void AddGrade(string subject, double grade)
         {
             if (!Subjects.ContainsKey(subject))
-                Subjects[subject] = new List<int>();
+                Subjects[subject] = new List<double>();
 
             Subjects[subject].Add(grade);
         }
